Add initial version of README file
This commit adds the initial version of the README file to the BuyIT project repository. The README provides instructions for getting started with the application, as well as information about its development status, building and running the application, and contributing to the project. This information will be helpful for users and developers who want to use or contribute to BuyIT.